Refactored code to render content when needed
