+++ src/js/components/TabContainer.js
@@ -7,7 +7,10 @@ import PostTweetUI from './PostTweetUI';
 const e = React.createElement;
 
 const Tab = (properties) => {
-    return e('button', {className: properties.className, onClick: properties.onClick, key: properties.key}, properties.tabName);
+    return e('button', {className: properties.className,
+                        onClick: properties.onClick,
+                        key: properties.key},
+                properties.tabName);
 }
 
 let tab;
@@ -24,13 +27,6 @@ class TabContainer extends React.Component {
 
     /* Change the content to be displayed */
     openTab(event, tabName) {
-//        if(this.props.children) {
-//            let tabContent = this.props.children.filter((i) => _.isEqual(i.key, tabName));
-//            this.setState({
-//                content: tabContent,
-//                currentTab: tabName
-//            });
-//        }
         let content;
         switch(tabName) {
             case "Home Timeline":
@@ -55,23 +51,7 @@ class TabContainer extends React.Component {
     }
 
     render() {
-        /* TODO: TEST THIS THOROUGHLY!! */
         return e('div', {className: "tabContainer"},
-//                    _.map(this.props.children, (content) => {
-//                        let append = "";
-//                        if(_.isEqual(this.state.currentTab, content.key)) {
-//                            append = " active";
-//                        }
-//
-//                        let tabProp = {
-//                            className: "tab" + append,
-//                            onClick: (event) => this.openTab(event, content.key),
-//                            key: content.key + " Tab",
-//                            tabName: content.key
-//                        }
-//
-//                        return Tab(tabProp);
-//                    }),
                 _.map(this.state.tabs, (tabName) => {
                     let append = "";
                     if(_.isEqual(this.state.currentTab, tabName)) {
