+++ src/scss/styles.scss
@@ -201,7 +201,6 @@ textarea {
         .postButtonWrapper {
             margin-top: 5px;
             text-align: right;
-            clear: left;
         }
 
         .charCounter {
@@ -214,7 +213,6 @@ textarea {
         .returnMessage {
             position: relative;
             display: inline;
-            clear: left;
             margin-right: 10px;
         }
